Feat: 엔티티 사망 및 피격 애니메이션 상태 통합 연동
- NetCharacter 단위에서 모든 생명체의 사망 시 물리 제어 및 Despawn 기능 중앙화
- NetPlayerController는 로컬 조작 차단에 집중하도록 책임 분리
- EntityAnimator에서 네트워크 의존성 없이 사망/피격 연출을 즉각 재생하도록 연동
- Rigidbody2D.velocity 등 레거시 API 교체 및 WaitForSeconds 캐싱 최적화 적용

+++ Assets/_Game/Scripts/Cores/Skills/Abilities/BasicAttackLogic.cs
@@ -17,7 +17,7 @@ namespace ProjectAI.Core.Skills.Abilities
         public ESkillType SkillType => ESkillType.BasicAttack;
 
         private NetworkObject projectilePrefab;
-        private double cooldown;
+        private float cooldown;
 
         public void Initialize(SkillManager manager)
         {
